refactor: shorten schemas import
