+++ api/server/middleware/denyRequest.js
@@ -1,7 +1,7 @@
 const crypto = require('crypto');
 const { saveMessage } = require('~/models');
 const { sendMessage, sendError } = require('~/server/utils');
-const { getResponseSender } = require('~/server/services/Endpoints/schemas');
+const { getResponseSender } = require('~/server/services/Endpoints');
 
 /**
  * Denies a request by sending an error message and optionally saves the user's message.
