+++ api/server/services/Endpoints/index.js
@@ -1,0 +1,5 @@
+const schemas = require('./schemas');
+
+module.exports = {
+  ...schemas,
+};
